requirements.txt
fastapi~=0.111.1
httpx~=0.27.0
redis~=5.0.7
starlette~=0.37.2
asyncio~=3.4.3
bson~=0.5.10
python-dotenv~=1.0.1